fix: stabilize i18n demo workflow and html entity decoding
