+++ .github/workflows/mr-test.yml
@@ -2,7 +2,25 @@ name: MR Checks
 on: [ pull_request ]
 
 jobs:
+  e2e-demo:
+    runs-on: ubuntu-22.04
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v6
+      - name: Set up Python 3.11
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+      - name: Install vx
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install vx
+      - name: Run demo workflow and demo script
+        run: |
+          vx just run-demo
+
   python-check:
+
     strategy:
       max-parallel: 4
       matrix:
@@ -18,7 +36,8 @@ jobs:
     runs-on: ${{ matrix.target.os }}
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
+
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
