+++ examples/locales/ko_KR/LC_MESSAGES/messages.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: TransX Demo 1.0\n"
 "Report-Msgid-Bugs-To: transx@example.com\n"
-"POT-Creation-Date: 2026-03-03 09:01\n"
+"POT-Creation-Date: 2026-03-07 12:03\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: ko_KR <LL@li.org>\n"
@@ -20,78 +20,78 @@ msgstr ""
 "Copyright-Holder: TransX Team\n"
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
-#: examples/demo.py:90
+#: demo.py:86
 msgctxt "greeting"
 msgid "Hello"
 msgstr "안녕하세요"
 
 
-#: examples/demo.py:91
+#: demo.py:87
 msgctxt "yes"
 msgid "Hello"
 msgstr "안녕하세요"
 
 
-#: examples/demo.py:93
+#: demo.py:90
 msgctxt "morning"
 msgid "Hello"
 msgstr "안녕하세요"
 
 
-#: examples/demo.py:95
-#: examples/demo.py:197
+#: demo.py:92
+#: demo.py:203
 msgid "Welcome {name}"
 msgstr "환영합니다 {name}"
 
 
-#: examples/demo.py:96
-#: examples/demo.py:196
+#: demo.py:93
+#: demo.py:202
 msgid "Hello"
 msgstr "안녕하세요"
 
 
-#: examples/demo.py:107
+#: demo.py:113
 msgid "Hello: $USERNAME"
 msgstr "안녕하세요: $USERNAME"
 
 
-#: examples/demo.py:108
+#: demo.py:114
 msgid "Hello: $${USERNAME} -> %USERNAME%"
 msgstr "안녕하세요: $${USERNAME} -&gt; %USERNAME%"
 
 
-#: examples/demo.py:134
+#: demo.py:140
 msgctxt "login"
 msgid "Hello"
 msgstr "안녕하세요"
 
 
-#: examples/demo.py:135
+#: demo.py:141
 msgid "Hello\\nWorld"
 msgstr "안녕하세요\\n세상"
 
 
-#: examples/demo.py:136
+#: demo.py:142
 msgid "Tab\\there"
 msgstr "Tab\\t여기"
 
 
-#: examples/demo.py:137
+#: demo.py:143
 msgid "Tab\\there!"
 msgstr "Tab\\t이 여기에 있습니다!"
 
 
-#: examples/demo.py:138
+#: demo.py:144
 msgid "Path to the file: \"C:\\Program Files\\MyApp\""
 msgstr "파일 경로: \"C:\\Program Files\\MyApp\""
 
 
-#: examples/demo.py:147
+#: demo.py:153
 msgid "如果出现错误，显示找不到 DCF_updateViewportList,或者找不到 CgAbBlastPanelOptChangeCallback，说明你的文件存在病毒，请运行一下打开优化工具，执行 remove unknown node and plugin 和 remove callback 两个按钮，然后保存一下文件"
 msgstr "오류가 발생하고 DCF_updateViewportList를 찾을 수 없다고 표시되거나 CgAbBlastPanelOptChangeCallback을 찾을 수 없다는 것은 파일에 바이러스가 있다는 의미입니다. 이를 실행하여 최적화 도구를 열고 알 수 없는 노드 및 플러그인 제거를 실행하고 콜백 버튼을 제거하십시오. 그런 다음 파일을 저장하십시오."
 
 
-#: examples/demo.py:198
+#: demo.py:204
 msgid "Settings"
 msgstr "설정"
