+++ vx.lock
@@ -1,0 +1,30 @@
+# vx.lock - Auto-generated, do not edit manually
+# Generated by vx 0.8.4
+
+version = 2
+
+[metadata]
+generated_at = "2026-03-07T03:30:38Z"
+vx_version = "0.8.4"
+platform = "x86_64-pc-windows-msvc"
+
+[tools.just]
+version = "1.46.0"
+source = "generic"
+resolved_from = "latest"
+ecosystem = "Generic"
+download_url = "https://github.com/casey/just/releases/download/1.46.0/just-1.46.0-x86_64-pc-windows-msvc.zip"
+
+[tools.python]
+version = "3.12.11"
+source = "python"
+resolved_from = "3.12"
+ecosystem = "Python"
+download_url = "https://github.com/astral-sh/python-build-standalone/releases/download/20250610/cpython-3.12.11+20250610-x86_64-pc-windows-msvc-install_only_stripped.tar.gz"
+
+[tools.uv]
+version = "0.10.9"
+source = "python"
+resolved_from = "latest"
+ecosystem = "Python"
+download_url = "https://github.com/astral-sh/uv/releases/download/0.10.9/uv-x86_64-pc-windows-msvc.zip"
